Begin refactor to Key Binds, cleanup tooltip mixin More will come after the mappings update
