--- library/gui/key_bindings/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	moduleName = "key_bindings"
-	version = "1.0.0"
-	library = "gui"
-	coreTestmodDependencies([
-			"qsl_base",
-			"lifecycle_events"
-	])
-}
-
-loom {
-	accessWidenerPath = file("src/main/resources/quilt_key_bindings.accesswidener")
-}
