--- library/gui/key_bindings/src/main/resources/fabric.mod.json
@@ -1,26 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "quilt_key_bindings",
-  "name": "Quilt Key Bindings API",
-  "version": "${version}",
-  "environment": "client",
-  "license": "Apache-2.0",
-  "icon": "assets/quilt_key_bindings/icon.png",
-  "contact": {
-    "homepage": "https://quiltmc.org",
-    "issues": "https://github.com/QuiltMC/quilt-standard-libraries/issues",
-    "sources": "https://github.com/QuiltMC/quilt-standard-libraries"
-  },
-  "authors": [
-    "QuiltMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.12",
-    "minecraft": ">=1.18"
-  },
-  "description": "Key bindings registration and utilities.",
-  "accessWidener": "quilt_key_bindings.accesswidener",
-  "mixins": [
-    "quilt_key_bindings.mixins.json"
-  ]
-}
