--- library/gui/key_bindings/src/main/resources/quilt_key_bindings.mixins.json
@@ -1,17 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.key.bindings.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "client": [
-    "client.EntryListWidgetAccessor",
-    "client.GameOptionsAccessor",
-    "client.GameOptionsMixin",
-    "client.KeyBindingAccessor",
-    "client.KeyBindingEntryMixin",
-    "client.KeyBindingMixin",
-    "client.KeyBindingsScreenMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
