--- library/gui/key_bindings/src/testmod/resources/assets/quilt_key_bindings_testmod/lang/en_us.json
@@ -1,7 +1,0 @@
-{
-    "key.qsl.category": "QSL Key Binding API Test",
-    "key.qsl.disable_key_bind": "Disable Key Bind",
-    "key.qsl.enable_key_bind": "Enable Key Bind",
-    "key.qsl.conflict_test": "Conflict Test",
-    "key.qsl.disabled_conflict_test": "Disabled Conflict Test"
-}
