--- library/gui/key_bindings/src/testmod/resources/fabric.mod.json
@@ -1,17 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "quilt_key_bindings_testmod",
-  "name": "Quilt Key Bindings API Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "quilt_key_bindings": "*"
-  },
-  "entrypoints": {
-    "client": [
-      "org.quiltmc.qsl.key.bindings.test.client.KeyBindingRegistryTestMod",
-      "org.quiltmc.qsl.key.bindings.test.client.KeyBindingsTestMod"
-    ]
-  }
-}
