+++ library/gui/key_binds/build.gradle
@@ -1,0 +1,17 @@
+plugins {
+	id("qsl.module")
+}
+
+qslModule {
+	moduleName = "key_binds"
+	version = "1.0.0"
+	library = "gui"
+	coreTestmodDependencies([
+			"qsl_base",
+			"lifecycle_events"
+	])
+}
+
+loom {
+	accessWidenerPath = file("src/main/resources/quilt_key_binds.accesswidener")
+}
