+++ library/gui/key_binds/src/main/java/org/quiltmc/qsl/key/binds/mixin/client/GameOptionsAccessor.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2021 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.key.binds.mixin.client;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.gen.Accessor;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+import net.minecraft.client.option.GameOptions;
+import net.minecraft.client.option.KeyBinding;
+
+@Environment(EnvType.CLIENT)
+@Mixin(GameOptions.class)
+public interface GameOptionsAccessor {
+	@Accessor
+	KeyBinding[] getAllKeys();
+
+	@Mutable
+	@Accessor
+	void setAllKeys(KeyBinding[] allKeys);
+}
