+++ library/gui/key_binds/src/main/java/org/quiltmc/qsl/key/binds/mixin/client/KeyBindingMixin.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2021 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.key.binds.mixin.client;
+
+import java.util.Map;
+
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
+import net.minecraft.client.option.KeyBinding;
+import net.minecraft.client.util.InputUtil;
+
+@Environment(EnvType.CLIENT)
+@Mixin(KeyBinding.class)
+public abstract class KeyBindingMixin {
+	@Shadow
+	@Mutable
+	@Final
+	private static Map<String, Integer> CATEGORY_ORDER_MAP;
+
+	@Inject(method = "<init>(Ljava/lang/String;Lnet/minecraft/client/util/InputUtil$Type;ILjava/lang/String;)V", at = @At("TAIL"))
+	private void addModdedCategory(String string, InputUtil.Type type, int i, String string2, CallbackInfo ci) {
+		if (!CATEGORY_ORDER_MAP.containsKey(string2)) {
+			CATEGORY_ORDER_MAP.put(string2, CATEGORY_ORDER_MAP.size() + 1);
+		}
+	}
+}
