+++ library/gui/key_binds/src/main/resources/assets/quilt_key_binds/icon.png
