+++ library/gui/key_binds/src/main/resources/quilt_key_binds.accesswidener
