+++ library/gui/key_binds/src/main/resources/quilt_key_binds.mixins.json
@@ -1,0 +1,17 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.key.binds.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "client.EntryListWidgetAccessor",
+    "client.GameOptionsAccessor",
+    "client.GameOptionsMixin",
+    "client.KeyBindingAccessor",
+    "client.KeyBindingEntryMixin",
+    "client.KeyBindingMixin",
+    "client.KeyBindingsScreenMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
