+++ library/gui/key_binds/src/testmod/resources/assets/quilt_key_binds_testmod/lang/en_us.json
@@ -1,0 +1,7 @@
+{
+    "key.qsl.category": "QSL Key Binds API Test",
+    "key.qsl.disable_key_bind": "Disable Key Bind",
+    "key.qsl.enable_key_bind": "Enable Key Bind",
+    "key.qsl.conflict_test": "Conflict Test",
+    "key.qsl.disabled_conflict_test": "Disabled Conflict Test"
+}
