+++ library/gui/key_binds/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,17 @@
+{
+  "schemaVersion": 1,
+  "id": "quilt_key_binds_testmod",
+  "name": "Quilt Key Binds API Test Mod",
+  "version": "1.0.0",
+  "environment": "client",
+  "license": "Apache-2.0",
+  "depends": {
+    "quilt_key_binds": "*"
+  },
+  "entrypoints": {
+    "client": [
+      "org.quiltmc.qsl.key.binds.test.client.KeyBindRegistryTestMod",
+      "org.quiltmc.qsl.key.binds.test.client.KeyBindsTestMod"
+    ]
+  }
+}
